requirements.txt
streamlit
pandas
numpy
scikit-learn
altair
python-dotenv
openai